requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
scipy>=1.11.0
requests>=2.31.0
